workers/pyproject.toml
[project]
name = "nirman-workers"
version = "1.0.0"
description = "Celery background workers for Nirman BaaS"
requires-python = ">=3.12"
dependencies = [
    "celery==5.3.6",
    "redis==5.0.3",
    "psycopg2-binary==2.9.9",
    "httpx==0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
